+++ inc/bms/devices/current_sensor/dev_isense.h
@@ -85,7 +85,7 @@ isense_status_t DEV_Isense_Get_Wait(isense_t *p_inst, uint32_t *p_out);
 
 isense_status_t DEV_Isense_Get_Gain(isense_t *p_inst, int32_t *p_out);
 
-isense_status_t DEV_Isense_Get_Val(isense_t *p_inst, int16_t *p_out);
+isense_status_t DEV_Isense_Get_Val(isense_t *p_inst, int32_t *p_out);
 
 isense_status_t DEV_Isense_Get_State(isense_t *p_inst, isense_state_t *p_out);
